+++ packages/terra-search-field/tests/wdio/search-field-spec.js
@@ -48,13 +48,14 @@ Terra.describeViewports('Search Field', ['medium'], () => {
   describe('With Placeholder', () => {
     before(() => browser.url('/#/raw/tests/terra-search-field/search-field/search-field-with-placeholder'));
 
-    Terra.it.validatesElement('placeholder');
+    // TODO: remove disabled color-contrast rule: https://github.com/cerner/terra-core/issues/2992
+    Terra.it.validatesElement('placeholder', { rules: { 'color-contrast': { enabled: false } } });
 
     it('should enter a search term', () => {
       browser.setValue('input', 'Lorem');
     });
 
-    Terra.it.matchesScreenshot('with text');
+    Terra.it.validatesElement('with text');
   });
 
   describe('With Default Value', () => {