+++ packages/terra-button-group/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Updated wdio test to support testing against firefox
+
 
 3.41.0 - (June 9, 2020)
 ------------------