+++ packages/terra-form-select/CHANGELOG.md
@@ -14,6 +14,8 @@ Unreleased
 ------------------
 ### Changed
 * Apply themes by context and directly include packaged themes.
+* Updated wdio test to support testing against firefox
+* Disabled wdio axe color-contrast testing on placeholder examples for IE testing. This will be re-enabled when https://github.com/cerner/terra-core/issues/2992 is resolved.
 
 6.5.0 - (May 19, 2020)
 ------------------