+++ packages/terra-paginator/CHANGELOG.md
@@ -14,6 +14,7 @@ Unreleased
 ------------------
 ### Changed
 * Apply themes by context and directly include packaged themes.
+* Updated wdio test to support testing against firefox
 
 2.50.0 - (May 26, 2020)
 ------------------